Added support for motion detection using the camera module, via picture class.
Two different images are run into the motion detection class, if they
are simmilar no motion will be detected. However if they are not similar
motion has been detected. 

The speed of the calculation depends highly of the CPU, not so fast on a
raspberry pi but if a cluster of them is put together the speed should
increase dramatically.